+++ openfisca_us/variables/expense/person.py
@@ -2,6 +2,16 @@ from openfisca_core.model_api import *
 from openfisca_us.entities import *
 
 
+class ProviderTypeGroup(Enum):
+    DCC_SACC = "Licenced/registered/permitted day care center; registered school-age child care"
+    FDC_GFDC = (
+        "Registered family day care homes; licensed group family day care"
+    )
+    LE_GC = "Legally exempt group child care programs"
+    LE_STD = "Informal child care standard rate"
+    LE_ENH = "Informal child care enhanced rate"
+
+
 class provider_type_group(Variable):
     value_type = Enum
     possible_values = ProviderTypeGroup
